pyproject.toml
[tool.poetry]
name = "chatbot-gradio"
version = "0.1.0"
description = ""
authors = ["Ali Noroozi <aliasghar.noroozi@gmail.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
gradio = "^4.26.0"
openai = "^1.17.1"
anthropic = "^0.25.2"
python-dotenv = "^1.0.1"
boto3 = "^1.34.91"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
